optimize api and fix bugs

+++ controllers/posts.js
@@ -60,6 +60,11 @@ export const updatePost = async (req, res) => {
 
     if(!mongoose.Types.ObjectId.isValid(_id))
         return res.status(404).send('No valid id');
+    
+    const currentPost = await PostMessage.findById(_id);
+
+    if(req.userId !== currentPost.creator)
+        return res.status(400).send('No permission');
 
     const updatedPost = await PostMessage.findByIdAndUpdate(_id, { ...post, _id }, { new: true });
     res.json(updatedPost);
@@ -70,6 +75,11 @@ export const deletePost = async (req, res) => {
 
     if(!mongoose.Types.ObjectId.isValid(id))
         return res.status(404).send('No valid id');
+
+    const currentPost = await PostMessage.findById(id);
+    
+    if(req.userId !== currentPost.creator)
+        return res.status(400).send('No permission');
 
     await PostMessage.findByIdAndRemove(id);
     res.json({ message: 'Post deleted successfully!' });
@@ -106,7 +116,7 @@ export const commentPost = async (req, res) => {
 
     post.comments.push(value);
 
-    const updatedPost = await PostMessage.findByIdAndUpdate(id, post, { new:true });
+    const updatedPost = await PostMessage.findByIdAndUpdate(id, post, { new: true });
 
     res.json(updatedPost);
 }
